Refactor tiles blocks sprites to drawable
