+++ foundry/gui/PlayerFrameEditor.py
@@ -4,6 +4,7 @@ from PySide6.QtGui import QCloseEvent
 from PySide6.QtWidgets import QHBoxLayout, QLayout, QStatusBar, QToolBar, QWidget
 
 from foundry import icon
+from foundry.core.drawable import SPRITE_SIZE, Sprite, SpriteGroup
 from foundry.core.geometry import Point
 from foundry.core.palette import PaletteGroup
 from foundry.core.player_animations import ANIMATION_WIDTH
@@ -12,7 +13,6 @@ from foundry.core.player_animations.util import (
     get_animations_palette_index,
     load_player_animation,
 )
-from foundry.core.sprites import SPRITE_SIZE, Sprite, SpriteGroup
 from foundry.gui.CustomChildWindow import CustomChildWindow
 from foundry.gui.Spinner import Spinner
 from foundry.gui.SpriteViewer import SpriteViewerController as SpriteViewer
